+++ graphql-gradle-plugin/src/main/java/com/graphql_java_generator/gradleplugin/GenerateClientCodeTask.java
@@ -135,6 +135,12 @@ public class GenerateClientCodeTask extends DefaultTask implements GenerateClien
 	}
 
 	@Override
+	@Input
+	public String getSpringBeanSuffix() {
+		return extension.getSpringBeanSuffix();
+	}
+
+	@Override
 	@OutputDirectory
 	public File getTargetClassFolder() {
 		return extension.getTargetClassFolder();