+++ graphql-gradle-plugin/src/main/java/com/graphql_java_generator/gradleplugin/GenerateGraphQLSchemaSpringConfiguration.java
@@ -26,12 +26,12 @@ import graphql.parser.Parser;
  */
 @Configuration
 // @Import({ JacksonAutoConfiguration.class })
-@ComponentScan(basePackageClasses = { DocumentParser.class, GraphqlUtils.class }, excludeFilters = {
-		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateClientCode.*"),
-		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateCode.*"),
-		@Filter(type = FilterType.REGEX, pattern = ".*\\.GeneratePojo.*"),
-		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateServerCode.*"),
-		@Filter(type = FilterType.REGEX, pattern = ".*\\.GraphQL.*") })
+@ComponentScan(basePackageClasses = { DocumentParser.class, GraphqlUtils.class }, //
+		excludeFilters = { @Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateClientCode.*"),
+				@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateCode.*"),
+				@Filter(type = FilterType.REGEX, pattern = ".*\\.GeneratePojo.*"),
+				@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateServerCode.*"),
+				@Filter(type = FilterType.REGEX, pattern = ".*\\.GraphQL.*") })
 public class GenerateGraphQLSchemaSpringConfiguration {
 
 	/**