+++ graphql-gradle-plugin/src/main/java/com/graphql_java_generator/gradleplugin/GeneratePojoTask.java
@@ -217,6 +217,12 @@ public class GeneratePojoTask extends DefaultTask implements GeneratePojoConfigu
 	}
 
 	@Override
+	@Input
+	public String getSpringBeanSuffix() {
+		return extension.getSpringBeanSuffix();
+	}
+
+	@Override
 	@OutputDirectory
 	public File getTargetClassFolder() {
 		return extension.getTargetClassFolder();