+++ graphql-gradle-plugin/src/main/java/com/graphql_java_generator/gradleplugin/GenerateServerCodeSpringConfiguration.java
@@ -26,11 +26,11 @@ import graphql.parser.Parser;
  */
 @Configuration
 // @Import({ JacksonAutoConfiguration.class })
-@ComponentScan(basePackageClasses = { DocumentParser.class, GraphqlUtils.class }, excludeFilters = {
-		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateClientCode.*"),
-		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateGraphQLSchema.*"),
-		@Filter(type = FilterType.REGEX, pattern = ".*\\.GeneratePojo.*"),
-		@Filter(type = FilterType.REGEX, pattern = ".*\\.GraphQL.*") })
+@ComponentScan(basePackageClasses = { DocumentParser.class, GraphqlUtils.class }, //
+		excludeFilters = { @Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateClientCode.*"),
+				@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateGraphQLSchema.*"),
+				@Filter(type = FilterType.REGEX, pattern = ".*\\.GeneratePojo.*"),
+				@Filter(type = FilterType.REGEX, pattern = ".*\\.GraphQL.*") })
 public class GenerateServerCodeSpringConfiguration {
 
 	/**