+++ graphql-gradle-plugin/src/main/java/com/graphql_java_generator/gradleplugin/GenerateServerCodeTask.java
@@ -169,6 +169,12 @@ public class GenerateServerCodeTask extends DefaultTask implements GenerateServe
 	}
 
 	@Override
+	@Input
+	public String getSpringBeanSuffix() {
+		return extension.getSpringBeanSuffix();
+	}
+
+	@Override
 	@OutputDirectory
 	public File getTargetClassFolder() {
 		return extension.getTargetClassFolder();