+++ graphql-gradle-plugin/src/main/java/com/graphql_java_generator/gradleplugin/GraphQLGenerateCodeTask.java
@@ -182,6 +182,12 @@ public class GraphQLGenerateCodeTask extends DefaultTask implements GraphQLConfi
 	}
 
 	@Override
+	@Input
+	public String getSpringBeanSuffix() {
+		return extension.getSpringBeanSuffix();
+	}
+
+	@Override
 	@OutputDirectory
 	public File getTargetClassFolder() {
 		return extension.getTargetClassFolder();